fix: resolve golangci-lint errcheck and unconvert violations
- Remove redundant uint32 cast in vector_getters.go (unconvert)
- Wrap defer Close() calls in require.NoError for errcheck compliance
  in enum_bench_test.go (conn, dkRows, Stmt) and types_test.go (rows x2)

+++ vector_getters.go
@@ -209,7 +209,7 @@ func (vec *vector) getEnum(rowIdx mapping.IdxT) string {
 	case TYPE_USMALLINT:
 		idx = uint32(getPrimitive[uint16](vec, rowIdx))
 	case TYPE_UINTEGER:
-		idx = uint32(getPrimitive[uint32](vec, rowIdx))
+		idx = getPrimitive[uint32](vec, rowIdx)
 	case TYPE_UBIGINT:
 		idx = uint32(getPrimitive[uint64](vec, rowIdx))
 	}
